补充数据概览

+++ src/kg4rd/extract/triple_supplement/data_parse.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 # Create Date: 2025/09/03
 # Author: wangtao <wangtao.cpu@gmail.com>
-# File Name: parse.py
+# File Name: data_parse.py
 # Description: 抽取到的三元组处理
 
 import json
@@ -11,26 +11,29 @@ import os
 import re
 import pandas as pd
 from tqdm import tqdm
+from pprint import pprint
+from typing import TypedDict
 sys.path.append('src/kg4rd')
 
 from synonyms.simple_search import simple_search
 
-# data prepare
-synonyms = json.load(open('src/kg4rd/extract/experimental_dmd/synonyms.json'))
 relation_type_name_map = json5.load(open('src/kg4rd/extract/relation_type_name_map.json5'))
 node_type_name_map = json5.load(open('src/kg4rd/extract/node_type_name_map.json5'))
 exists_edges = pd.read_csv('src/kg4rd/kg/kg.csv')
 exists_nodes = pd.read_csv('src/kg4rd/kg/nodes.csv')
 
-existing_edges_set = set(
-    (row['relation'], str(row['x_id']), str(row['y_id']), row['x_type'], row['y_type'])
-    for _, row in exists_edges.iterrows()
-)
+existing_edges_set = set(zip(
+    exists_edges['relation'],
+    exists_edges['x_id'].astype(str),
+    exists_edges['y_id'].astype(str),
+    exists_edges['x_type'],
+    exists_edges['y_type'],
+))
 
-existing_nodes_set = set(
-    (row['node_id'], row['node_type'])
-    for _, row in exists_nodes.iterrows()
-)
+existing_nodes_set = set(zip(
+    exists_nodes['node_id'],
+    exists_nodes['node_type'],
+))
 
 result_path = 'data/data_abstract/result'
 
@@ -39,33 +42,36 @@ def parse(data: list[dict], name: str, save_df: bool = True):
     approved_triples = []
     approved_triples_node_exist = []
 
-    for result in tqdm(data):
+    for result in tqdm(data, position=1, leave=False):
         relation_choices = result['relation_choices']
-        for triple in result['extracted_relations']:  # 获取到的三元组
+        for item in result['extracted_relations']:  # 获取到的三元组
             
-            subject, object_, relation = triple['subject'], triple['object'], triple['predicate']  # SOP
-            
-            original_triples.append({
+            subject, object_, relation, uid = item['subject'], item['object'], item['predicate'], item['uid']  # SOP
+            triple = {
                 'subject': subject,
                 'object': object_,
                 'relation': relation,
-                'uid': triple['uid'],
+                'uid': uid,
                 'status': ''
-            })
-        
+            }
+            
+            original_triples.append(triple)
+
             rs = re.sub(r'\s*\([^)]*\)', '', relation)
             rs = [r.strip() for r in rs.split('-')]
+            if len(rs) != 2:
+                continue
             
-            subject_type = node_type_name_map.get(rs[0])
-            object_type = node_type_name_map.get(rs[1])
+            subject_type = node_type_name_map.get(rs[0], '')
+            object_type = node_type_name_map.get(rs[1], '')
             
             subject_ret = simple_search(subject, subject_type)  # 名称和类型可能都有问题所以写作 synonym/id
             object_ret = simple_search(object_, object_type)
             
             if subject_ret is None and object_ret is None:
-                original_triples[-1]['status'] = 'TWO_SIDE_NO_SYNONYM/ID'
+                triple['status'] = 'TWO_SIDE_NO_SYNONYM/ID'
             elif subject_ret is None or object_ret is None:
-                original_triples[-1]['status'] = 'ONLY_ONE_SIDE_SYNONYM/ID'
+                triple['status'] = 'ONLY_ONE_SIDE_SYNONYM/ID'
             elif object_ret is not None and object_ret is not None:
                 
                 subject_id, subject_preferred_name, subject_preferred_name_score = subject_ret
@@ -73,12 +79,13 @@ def parse(data: list[dict], name: str, save_df: bool = True):
 
                 
                 if relation not in relation_choices:
-                    original_triples[-1]['status'] = 'RELATION_TYPE_NOT_IN_CHOICES'
+                    triple['status'] = 'RELATION_TYPE_NOT_IN_CHOICES'
                 else:
+                    relation = relation_type_name_map.get(relation)
                     if (relation, subject_id, object_id, subject_type, object_type) in existing_edges_set:
-                        original_triples[-1]['status'] = 'KG_ALREADY_EXISTS'
+                        triple['status'] = 'KG_ALREADY_EXISTS'
                     else:
-                        original_triples[-1]['status'] = 'APPROVED'
+                        triple['status'] = 'APPROVED'
                         at = {
                                 'relation': relation,
                                 'x_type': subject_type,
@@ -92,14 +99,14 @@ def parse(data: list[dict], name: str, save_df: bool = True):
                                 'y_preferred_name_score': object_preferred_name_score,
                             }
                         if (subject_id, subject_type) in existing_nodes_set and (object_id, object_type) in existing_nodes_set:
-                            original_triples[-1]['status'] = 'APPROVED(NODE_ALL_EXISTS)'
+                            triple['status'] = 'APPROVED(NODE_ALL_EXISTS)'
                             approved_triples_node_exist.append(at)
                         approved_triples.append(at)
-
-    df_original_triples = pd.DataFrame(original_triples)
+    
+    df_original_triples = pd.DataFrame(original_triples, columns=['subject', 'object', 'relation', 'uid', 'status'])   # pyright: ignore[reportArgumentType]
     df_approved_triples = pd.DataFrame(approved_triples)
     df_approved_triples_node_exist = pd.DataFrame(approved_triples_node_exist)
-    
+
     overview = {
         'abstract count': len(data),  # 摘要数量
         'extracted triples': len(df_original_triples),  # 抽取到的三元组数量
@@ -111,24 +118,25 @@ def parse(data: list[dict], name: str, save_df: bool = True):
         'approved triples node exist/before deduplication': len(df_approved_triples_node_exist),  # 可用的三元组数量(双侧节点已存在)(去重前)
     }
     
-    # 对可用的三元组去重 (包括严格双侧实体存在的部分)
-    df_approved_triples.drop_duplicates(subset=['relation', 'x_id', 'y_id', 'x_type', 'y_type'], inplace=True, keep='first')
-    df_approved_triples_node_exist.drop_duplicates(subset=['relation', 'x_id', 'y_id', 'x_type', 'y_type'], inplace=True, keep='first')
-
-    overview.update({
-        'approved triples': len(df_approved_triples),  # 可用的三元组数量(去重后)
-        'approved triples node exist': len(df_approved_triples_node_exist),  # 可用的三元组数量(双侧节点已存在)(去重后)
-    })
-    
     if save_df:
-        df_original_triples.to_csv(os.path.join('data/data_abstract/original_triples', f'{name}.csv'), index=False)
-        df_approved_triples.to_csv(os.path.join('data/data_abstract/approved_triples', f'{name}.csv'), index=False)
-        df_approved_triples_node_exist.to_csv(os.path.join('data/data_abstract/approved_triples_node_exist', f'{name}.csv'), index=False)  # 最后可用
+        os.makedirs('data/data_abstract/original_triples', exist_ok=True)
+        os.makedirs('data/data_abstract/approved_triples', exist_ok=True)
+        os.makedirs('data/data_abstract/approved_triples_node_exist', exist_ok=True)
+        if df_original_triples.shape[0] > 0:
+            df_original_triples.to_csv(os.path.join('data/data_abstract/original_triples', f'{name}.csv'), index=False)
+        if df_approved_triples.shape[0] > 0:
+            df_approved_triples.to_csv(os.path.join('data/data_abstract/approved_triples', f'{name}.csv'), index=False)
+        if df_approved_triples_node_exist.shape[0] > 0:
+            df_approved_triples_node_exist.to_csv(os.path.join('data/data_abstract/approved_triples_node_exist', f'{name}.csv'), index=False)  # 最后可用
     
     return overview
-    
-for file in os.listdir(result_path):
+
+overview = {}
+for file in tqdm(os.listdir(result_path), position=0):
     with open(os.path.join(result_path, file), 'r', encoding='utf-8') as f:
         data = json.load(f)
-        overview = parse(data, file, save_df=True)
-        ...
+        for k, v in parse(data, file.split('.')[0], save_df=True).items():
+            overview[k] = overview.get(k, 0) + v  # 统计更新
+
+with open('data/data_abstract/overview.json', 'w', encoding='utf-8') as f:
+    json.dump(overview, f, ensure_ascii=False, indent=4)
